requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
ultralytics==8.1.25
Pillow==10.2.0
PyMuPDF==1.23.21
numpy==1.26.4
easyocr==1.7.1
opencv-python==4.9.0.80
boto3==1.34.119
botocore==1.34.119
onnxruntime==1.17.3 
